Split embedder into impl + re-export so tests can bypass module mocks
Bun's mock.module() is process-wide. test/chat/agent.test.ts and
test/worker/prompt.test.ts stub src/context/embedder.ts, and on CI the
mock leaks into test/context/embedder.test.ts — its "real embedder"
assertions then fail because embed() is the zero-vector stub.

Move the implementation to src/context/embedder-impl.ts and turn
embedder.ts into a thin re-export. Other consumers still import from
embedder.ts (so existing mocks still work), but embedder.test.ts now
imports from embedder-impl.ts — a path that nothing mocks.
